MONITOR/HTML: Sync behavior of expired values with Qt version
The Qt version of the status monitor has an option 'noexpired' which
allows to display 'n/a' for the expired values. This option is now
available for the HTML version as well and behaves like in the Qt
version. So both version of monitors should display the same if
configured identically.

Change-Id: Id5edb74481d7d7c2e8c71a43d2813f182a24e161
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/22217
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/22260
Tested-by: Christian Felder <c.felder@fz-juelich.de>
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>