Change to Kafka Consumer
According to the Kafka docs, querying the end offsets will hang
indefinitely if the partition is unavailable. This means that
NICOS will not report this.
This changeset will avoid this, so NICOS will show a disconnected
message soon after the partition disappears. It will also respond
quicker to the FW being stopped.

Change-Id: I4f9a13783bf25cb6ca7fce5cb775f8e3bf89424b
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/22273
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>