watchdog: allow defining conditions in setups and load them automatically
Providing conditions together with global sample environment setups will
actually make the watchdog consistently useful for the SE group.

It is envisioned that these conditions use a notifier type like 'se',
and the addresses that these notifications goes to is configured in the
instrument specific notifier setup.

Even in non-SE setups this feature provides modularity, and together
with further simplifications a very quick way for instrument people to
add conditions.

Change-Id: I17ac00c400e6c771229afab4a7ac5eb38949647a
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/22300
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>