GUI commandlets: sorting and deregistration
Every category gets a priority. (default 100)
Every cmdlet gets a priority. (default 100)

Same priorities are sorted by insertion order.

The lower the priority value the higher/further
left the commandlet/category will be listed.

Adding an already existing commandlet with a
different priority will move the commandlet to
that priority. Same goes for categories.

Quickaccess to some commands is not needed/wanted
on certain instruments. Thus deregistration of the
corresponding commandlets is now an option.

Change-Id: I68782c7bfd01baeb0eeb75f1c995133884b72f6d

GUI cmdlets: allow deregistration of default cmdlets

Quickaccess to some commands is not needed/wanted on certain
instruments. Thus deregistration of the corresponding commandlets
is now an option.

Change-Id: Iadf894c4c0fe8a8ff8766da52ae4ce1662a6de2d
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/30240
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>