GUI: Add a new function to compose panels using VBoxLayout and HBoxLayout
When the GUI is composed using v(h)split, the user can resize the panels. If
the GUI is then shout down and reopened, the panels will have the size they had when the GUI
was closed.
In order to prevent the user to resize the panels inside the GUI, the new composition
functions "vbox", "hbox" are added. The subpanels are created inside a widget with V(H)BoxLayout.

Change-Id: Ife473ff732d231490a45e1f09cdfed31fa09e5f3
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/22658
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Matt Clarke <matt.clarke@esss.se >