SECoP support
Depending on the use case, devices are generated dynamically after
a successful connection, or they might be explicitely mentioned
in a setup.

The introduction of dynamically generated devices has some implications:
We can not register these devices in session._setup_info, as they would
have be overwritten on an further AddSetup command. They are registered
separately in session.dynamic_devices and are merged to the result
in session.getSetupInfo() instead. When the SEC node creates the devices
it assigns them to the same setup as the SEC node itself.

Added feature:
- make dry run work

Change-Id: I9b5d6759de92bbf77f600589e6376fba55338ca1
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/22844
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>