ESS: Added request/response functionality for just-bin-it
When a command is sent to just-bin-it it will respond with an ACK or ERR
message to say whether it has accepted the command.
If the histogramming starts but then subsequently fails for some reason the
last histogram message will flag the error and the reason.
See just-bin-it README for more information.

Also refactored "slave_mode" so it no longer needs an extra parameter, but
should still work how Mark needs it to.

To test:
Requires the very latest version of just-bin-it off master and that to be run
with the --response-topic parameter defined. The response topic should not be
the same as the command topic.

Likewise for NICOS, in the setup file for just-bin-it the (same) response topic
also needs to be defined.

Note: now uses the streaming-data-types module, for deserialising hs00 messages.

Change-Id: I37a85c444248cff43f2869f813f8de32fb82d56c
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/23052
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Matt Clarke <matt.clarke@esss.se>