EPICS: add AreaDetector image channel
This commit allows to display images acquired with the EPICS AreaDetector module.
The device is aware of informations s.t. image size, binning, zero point and data type.
In general the PV prefix for the data array can be different from the PV prefix for the camera setup.
Note that the EPICS ca library must be explicitely initialized, to make sure that the memory for the ArrayData PV is allocated.

Change-Id: I35fa4ab93ca97adda81a41ef476001ef1e82a115
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/23108
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Matt Clarke <matt.clarke@esss.se >