ESS/GUI: select which detector to display in live view
Currently, if more 'session.experiment.detectors' contains more than
one detectors, the live view tries to display all the images as soon
as the livedata arrives. The result is that the last detector
processed replace the previous ones.
This commit introduces a new Panel that displays a preview of all the
detectors set for the experiment and one of them as the main live
image.

Change-Id: I963f6422dd952ac8b1489a65961e5796e2725be3
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/23139
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Matt Clarke <matt.clarke@esss.se>