ESS: Remove connected "tick" from menu
This removes the tick that appears on the main menu when connected.
As the menu item changes from "Connect to server" to "Disconnect"
and the GUI itself shows it is is connected/disconnected the tick is
not necessary.

Also, we changed some camelCase names so this PR is a bit larger than
it could have been - sorry!

Change-Id: I3ce10dda1058b07474a26cc6aa6de804a47f5b3b
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/23988
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Ümit Hardal <umit.hardal@ess.eu>
Reviewed-by: Michele Brambilla <michele.brambilla@psi.ch>
Reviewed-by: Matt Clarke <matt.clarke@esss.se>