ESS: Refactor initialisation of toolbar out of from __init__ and override it in the ESS child
This refactors the initialization of the toolbar out of from the init constructor of the parent class and wraps it in a private function
which can be properly override in the ESS child without chaning any functionality in core Nicos. We change the connection status label from a LED
(which is green when connected and gray when disconnected) to the texts "Connected" and "Disconnected" and moved it to the write of the GUI in accordance
with the wishes of ESS instrument scienctists.

Change-Id: I6a363d89afb0f0920308788bc461ed62fc10051e
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/24024
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Matt Clarke <matt.clarke@esss.se>