GUI: Allow customisation of disconnect and view-only behaviour
With the newer GUI, the setups panel is always visible, so the controls
need to be disabled when disconnected or in view-only mode. Also the
"close" button is no longer required.
To do this required adding setViewOnly to the base setups class and
modifying the __init__ so the button behaviour could be overridden.

Change-Id: I0432769714939e7c5b086d4cf50a082be85ceb59
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/24200
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Matt Clarke <matt.clarke@esss.se>