ESS: disable toggling expert mode when disconnected
Unlike in https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/24246 I
don't think we do this via the UI file as the "connect" item is not
checkable in the newer gui.
Instead we can do it in code. We will also trial switching off expert
mode on disconnection.

Change-Id: Ie1b7f3e4281bce6d844217ea54eb581363177002
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/24252
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Ümit Hardal <umit.hardal@ess.eu>
Reviewed-by: Matt Clarke <matt.clarke@esss.se>