EPICS: replace HasSwitchPv with HasDisablePv
The non-EPICS implementation of a device that can be enabled/disabled
(whatever this may mean) is achieved with the use of CanDisable mixin.
The EPICS devices used to have a HasSwitchPv that allow to switch on and off.
This patch replaces the HasSwitchPv mixin with a HasDisablePv that inherits
from CanDisable.
The advantage is a more broad meaning of the switchpvs and the possibility to
enable/disable an EPICS device from the DeviceDialog.

Change-Id: If13ef972cea3bdc088d17a6130ff8cefa980b6fe
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/24527
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Ümit Hardal <umit.hardal@ess.eu>
Reviewed-by: Matt Clarke <matt.clarke@esss.se>