Axis: Set motor position to target if no coder
If the axis has no coder the value of the motor should be set to
the last target value if the target differs from the current read
value.

This helps to avoid references if the hardware was switched off/on
but there wasn't any move. It's the responsibility of the users to
verify it. A warning about this action will be given.

Fixes: #4108

Change-Id: I063ceaabf27b74bbb623b014c6a5a4d58c817294
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/24666
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>