tango: ImageChannel readout should block on final readout
The current implementation may leads to intermediate readout results
when reading an tango.ImageChannel on quality final because we make
the read out call regardless of the current status immediately after
calling det.finish(). A tango detector may report it is still in
progress of reading out the detector but still returns the previous
image on read value. Blocking on tango layer does not make sense here
as this can raise timeout errors when readout takes longer than 5s.

Basically this can be lifted into Measruable.readResult as well but
I am not sure whether we want to introduce this in general for all
other middlewares, e.g. EPICS.

Change-Id: Ib52439b22875c87e9f26007bf6dfbef142b1a6a1
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/24796
Tested-by: Christian Felder <c.felder@fz-juelich.de>
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>