gui/devices: show failed-to-create devices in the list as "error"
There is a pretty arbitrary difference in how different device error
conditions are shown in the device panel:

* if a device fails to create (e.g. Tango server is down), it is
  not shown at all and it is quite easy to miss when people don't
  check the error messages at setup load.

* if a device fails after creation (e.g. Tango server is down),
  it is very visible shown as "error".

From the user's standpoint it is not clear why there is a differnce
in behavior here, and it can be challenging to explain it.

This change tries to fix that by the daemon telling the device list
when devices fail to create, and the device list showing them as
"error", and offering a menu item to try to re-create them.

Change-Id: Ib238f73e0cae76a4119912eb02798c1b61ec4b72
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/24957
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>