Adding save functionality to histogram plot
Added code to save histogram plot.
Updates are done both to ui file and python files in nicos core and
nicos ess.
In almost all cases however there is an issue that seems to be
related to the gr package. The problem is that graph axes are not
saved properly for all file formats except for the pdf file format.

Change-Id: I7e2889c0478c56a700b1507c18da09c2d02c7e2b
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/24968
Reviewed-by: Matt Clarke <matt.clarke@esss.se>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>