guisupport: Fix Z scale calculation for the 2D widget
Change the scale of the Z values for the 2D scans. It seems that
the order of performing the scale calculation is a little
problematic with the numpy array. Instead of
  arr = val + scale * array / divider
its better to make it like this
  arr = val + scale / divder * array
Under some condition the first way does not lead to the correct
results but in second case the values are correct calculated.

Addionally the number of operations on the arrays will be reduced
by one.

Change-Id: I380d93448cd1c94f9b9ccf4d0813451e1a63a049
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/25177
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>