GUI: Reduce the scope of shortcut keys for copy, cut and paste
Using ctrl+c, ctrl+v, ctrl+x on a table on a custom panel gives a warning about the shortcuts being ambiguous and they are ignored.
This patch fixes the issue by reducing the context scope of the pre-existing short-cuts on other panels.
It doesn't affect the existing behaviour of the changed panels (tested on the KWS and standard GUIs).


Change-Id: I74946071c704857a939623316c50384b4e1fb2a6
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/25490
Reviewed-by: Christian Felder <c.felder@fz-juelich.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>