Make more flexible the creation of NexusFileSink file name.
The user might want the path of the file output dir to be
(auto)configured at runtime, e.i. using the experiment
name (this is the case with the AmorNexusFileSink).
With this patch the default behaviour of the NexusFileWriterSink
is to use the 'file_output_dir' parameter, but subclasses
can easily refine their own rule.

Change-Id: Ib35f54a5a728fd96bde25515d4a6077c7e5f6b96
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/25972
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Mark Koennecke <mark.koennecke@psi.ch>
Reviewed-by: Matt Clarke <matt.clarke@esss.se>