GUI: MultiLiveDataPanel recreate previews on dimension change
If we switch a just-bin-it image from a 2-D plot to a 1-D plot at run time then the preview is corrupted because it is trying to plot 1-D data on a 2-D plot widget. The same happens for the opposite change.
This patch checks to see if the dimensionality has changed and recreates the previews, if required.

Change-Id: I639254fca50da3b9f8f99f7d4a96e65a78910566
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/26165
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Michele Brambilla <michele.brambilla@psi.ch>
Reviewed-by: Matt Clarke <matt.clarke@esss.se>