EPICS: option to ignore stop for MappedMoveable
The usual behaviour for a moveable when "stop" is actioned is to request the device to move to its current position (in effect stopping). However, for a MappedMoveable this means it can re-send the current value, for example: "ON", which may or may not cause the device to do something.
This patch makes this behaviour opt-in.

Change-Id: I06b1aa12862465045b01011790d6b80b99e0e739
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/26254
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Matt Clarke <matt.clarke@esss.se>