NEXUS: make chunk_size configurable in template
The size of the NeXus file generates is too big w.r.t its content.
This is caused by the default chunk_size configured in the filewriter.
With this change the EventStream accepts the 'chunk_size' parameter,
that allows to configure the corresponding entry in the command.
In principle, the parameter can be configured on each KafkaStream,
but it's fully implmemented only in EventStream.

Change-Id: Ibd53717d7800a96643190c19b60d6e8a07040f17
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/26428
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Matt Clarke <matt.clarke@esss.se>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Mark Koennecke <mark.koennecke@psi.ch>
Reviewed-by: Michele Brambilla <michele.brambilla@psi.ch>