EPICS: add devices that use the p4p library for PVA
This patch adds versions of the EPICS devices that use the p4p library to support the PVA protocol.
Compared to pvaPy, p4p is generally nicer to use and it can also be used concurrently with PyEpics.
The implementation includes both monitor and non-monitor support which can be toggled via a parameter.

Change-Id: I44f1106a6c5deecd651e2cc4fb3ac72af951586d
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/26681
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Michele Brambilla <michele.brambilla@psi.ch>
Reviewed-by: Matt Clarke <matt.clarke@esss.se>