EPICS: allow device classes to be used for both protocols
Add a parameter to EpicsDevice to switch between PVA and CA.
This would allow the derived classes such as EpicsMotor to be used with
both protocols.

Why caproto? It is a pure Python implementation of CA and it does not
require the complicated setup that pyepics does regarding threads (see
pyepics.py).
Downsides are that it isn't as widely used as pyepics, but we've been
using it on other projects and NSLS-2 use it extensively.

Change-Id: Ib478bd426fb8d62efc6c8cf326e223ea47bc7473
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/26820
Reviewed-by: Matt Clarke <matt.clarke@esss.se>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>