NeXus: close file after update
It turned out the the call to flush() was not enough to get the data to appear on disk after an
update. The reason could be the buffering happening in linux. This change closes the file after an
update. Now the data appears on disk and can be processed with other software while the measurement
is still running.

Change-Id: Ib6c83632ecd58e6750769d1ca405e199e14d5d34
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/26826
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Mark Koennecke <mark.koennecke@psi.ch>