history/monitor: support real expressions for key specification
Currently, for specifying keys for history view and in display widgets
(DisplayLabel and TrendPlot, used for the monitor) we only allow a
very restricted pseudo-syntax, i.e. key[index]*X+Y.

This patch extends the syntax to full Python expressions, including
some usual math functions also present in the NICOS namespace.
There is still a restriction that only one cache key can be used
per expression.

Now finally there is no need to remember to use "a*0.001" instead
of "a/1000", or that parentheses are not supported.

As a bonus, this also enables scale/offset in all places; it was
previously not heeded except in the history plotter.

Change-Id: Id836007700d94fac736300adee2b53581a5ac67a
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/26852
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>