ESS: Generic counter for an EPICS value
Counter that sums a PV value until a certain target is reached.
Use case is something like summing the charge per pulse and stopping
counting when a certain value is met, e.g. count(charge=30).

In some ways it is similar to change #25991 but there won't be counter
hardware we can zero, instead we will only be able to consume what the
accelerator sends us.

Change-Id: If1562c3fff6374d7e2ca5cd90cf4e97812fd95d1
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/27388
Reviewed-by: Ebdad Kamil <ebad.kamil@ess.eu>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Matt Clarke <matt.clarke@esss.se>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>