GUI: add context menu to history views
Adds a context menu to the views in the list.

Without this change, on Mac, right-clicking to delete a view somehow
means that the "New" and other toolbar buttons no longer work. This does
not happen on Linux.

Generally, when there is more than one view right-clicking deletes two
views rather than one this also fixes that issue.

Change-Id: I080d9dc26cda3cedf1d4e70e06dea28202f9ec5c
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/27660
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Tested-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Matt Clarke <matt.clarke@esss.se>