Test: Create structure for facility specific tests
The facility specific test setups have to be stored in a 'setups'
directory in the facility specific test directory of the test directory:

Example: The setups for the general MLZ tests (not instrument specific)
have to be stored in 'nicos_mlz/test/setups'

For test setups for instrument specific tests, the files should be
stored in a directory like nicos_'facility'/'instrument'/test/setups.

Example: MIRA specific test setups should be store in
'nicos_mlz/mira/test/setups'.

This restructuring should help to avoid very long common setup file
names containing hints to the facility and/or instrument name.

The names of the setup files can be choosen as for the instruments
itself, inside the facility/instrument/test the setup file names are
uniq, but in various instruments the same name can be reused.

Change-Id: Ib52389845a043ad04f2ef37e1a22ec5988a7f2c9
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/27987
Reviewed-by: Stefan Mathis <stefan.mathis@psi.ch>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>