GUI: Add default connections in GUI config
In the gui config the options may be used to add default connections
presets in the 'connection_presets' dictionairy. The key is the name of
the connection and the value is the connection string in the same format
as the command line parameter. As the default the 'viewonly' of this
connections will be set true.

The connection presets override existing settings during the startup.

Fixes: #3449

Change-Id: I26858c75f8e56bde94f431b0de3c74f6117302b0
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/28254
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>