SANS1: Add 'valuename' and 'unit' of IEEE to BerSANS file
For the application of the generated BerSANS files it maybe very helpful
to have the used device or device parameter of the IEEE devices in the
data file itself. The will be stored in the '%Comment' section of the
file as (IEEE device name)_valuename and (IEEE device name)_unit.

Change-Id: I6cd6c7795823c41f0484deabd7670f6c2c2ab669
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/28261
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>