LDAPAuth: Allow multiple LDAP servers
In case one have more than one dedicated LDAP server it's helpful to
have a series of failover servers. The 'uri' parameter has been changed
to a list of URI. The first of the servers in the list which responds
wins the race. So one may control the priority of the LDAP servers.

Change-Id: I5f36d8092f252164effb27e58bd37c7fe9a58756
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/28263
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>