core/help: adds additional help topics
New help topics are supposed to be placed as a dictionary in setup file
of an instrument or/and to setup files of devices within an instrument.

These help topics are loaded to the session when setups are loaded,
from where they are accessed by help generator.

New link tag "topic:" is added to browse these topics on help page.

New help topics (if they exhist) are added to the beginning of the index page
and they are clickable.

Tests are updated to check one new help topic, "topic:" link tag, and
RST functionality.

New method added to check-setups, in order to search for help topics
duplicates and return an error, if any.

Fixes: #4189

Change-Id: I91866ebeabde0739b7975c3962c17aae4d2e25c4
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/28415
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>