ESS: only delete 'valid' rows
On deleting rows from the bottom of the table, qt will retain the
indices even though they are invalid. Prior to this patch pressing
delete again would try to delete these already deleted rows.

Also corrects the behaviour of adding a row to an empty table and adding
a row above the current one.

Change-Id: I24dca0b74b0df1d703b147141aa0435f136caa2f
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/28463
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Matt Clarke <matt.clarke@esss.se>