gui/history: fixed memory leak when history panel is closed
When there are active views in the history panel, then if panel is closed
and views were not deleted beforehand, they remain in the memory. Timers
continue to tick and query the cache. Restoring the view, creates new view
and the previous one is still in the memory.
Cleanup() method properly removes the view, then it is still possible to
restore it in next history session.

Change-Id: Ib103091d4edad217b7077ba50d0c79c51e7afd7d
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/28474
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Konstantin Kholostov <k.kholostov@fz-juelich.de>