gui/history: long-time cache queries are executed properly
Fixes: #4418
There are two timeouts that limit cache query: 1) client opens socket
for 30 secs, 2) cache request timeout is 10 secs. Since it is not
possible to predict how much time is necessary to read the cache,
it is proposed here to read cache in sequence if requested period is
longer than 1 day.

Change-Id: I200d2159a79e255daff4554f2e60b3c1d907cd76
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/28500
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Konstantin Kholostov <k.kholostov@fz-juelich.de>