flowui: avoid multi-liveview segfault
Moving the splitter between the liveview and the previews causes a
segfault when the combined height of the previews cross the value at
which the vertical scrollbar should appear or disappear.
I suspect this is because the size is momentarily incompatible with the
visibility of the scrollbar.

Simple fix: always display the scrollbar!

Change-Id: I21e6d91b59c0eda77519f064cc6191794944885d
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/28973
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Matt Clarke <matt.clarke@ess.eu>