SECoP: avoid pickling the data
On the client side, the import of the secop package should
not be needed. In the previous version, the setup_info stored
on the connection device contained data which needs pickle
(extended nicos data types as classes). Reworked to use the native
SECoP datainfo instead, which is just dicts, lists, numbers
and strings.

Fixes: #4575

Change-Id: I3f2c6615ba7d9a437e6ec85ff31b19caaf843c69
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/28984
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Markus Zolliker <markus.zolliker@psi.ch>