gui: minor bug in auth dialog
If one has multiple presets and the last preset loaded was not the
first one in the list then if the first preset is selected next then
the username, view-only and expert settings are not updated because
the combobox index is already 0.

Using setCurrentIndex to set the last preset is one way to fix this.

Change-Id: I30ba92c5ebc8ad2133c374962b1568dd76f8e94e
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/29081
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Matt Clarke <matt.clarke@ess.eu>