SINQ: special version of JustBinIt
The default integration of just-bin-it only allows to count on an
(internal) timer. This is not compatible with the default user case
at SINQ, where users count on monitor events.

With this patch just-bin-it can be used as a simple image channel,
and the default Detector device deals with all the channels properly.

Change-Id: Ibb09d1b1cb686ae4779201e4bcbd4d26509e7a1a
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/29103
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Michele Brambilla <michele.brambilla@psi.ch>