Detectors: Add preset query function
Add an interface function to query the current used presets in the
Measurables/Detectors.

This can be used to query the current set presets for this devices if
some of the measuring commands like 'scan' or 'count' rely on the last
set presets (in other words the command will not have any presets).

The scans and counts now using this feature to ensure the data sets have
presets even if the command itself don't get them from user input.

Fixes: #4609

Change-Id: I570e2608b6dc83bee90236ab8879b534c1a95081
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/29145
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>