GUI: Add option to configure ssh tunnel interactively
Since the command line option to configure the ssh tunnel is not very
comfortable for standard users an option to configure the ssh tunnel via
the graphical setup would be quiet useful. It's helpful as well for the
windows version.

The sshtunnel package requirements are increased to fix some problems
with finishing the client.

Change-Id: Iacdf75acd0696b6afc7a4077a3059228a4ac7bcb
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/29235
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Bjoern Pedersen <bjoern.pedersen@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>