TOFTOF: Make the safety status device less chatty
The safety bit state information shouldn't be given all the time, if a
count is startet. It gives not very helpful to the users. If really
needed in the script the output can be enabled via setting the loglevel
of the device to 'debug'.

It cleans up the output of scans as well.

Change-Id: Icf27668046f7e855e3eef56a44c69b17c8d29ad5
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/29361
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>