Monitor/HTML: Fix expire setting
In some rare cases the expire setting of a value is wrong due to the
fact, that the last key submit is the format. In that case the expire
value will be set to False, which is not in all cases the right choice

At sans1 we had the case that the keys were sent in the following order

'unit', 'value', 'status', 'fixed', 'fmtstr'. The 'status' and 'value'
were expired but the 'fmtstr' reverted the 'expire' by setting it to
False in any case.

Change-Id: Id2731b77f6de3eb80f3096dd62d8f3b9ee0926e1
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/29413
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>