cache: InfluxDB cache back-end
Influxdb_client module is used to communicate with InfluxDB2.0 instance.
Nicos values are stored in _fields, expiration is marked as _tag, this
is done because when data is queried, tag is already presented in every
InfluxDB record. Each table is returned in separate tables, so if
expired is set in _fields instead of _tag, this requires extra code to
handle. Nicos kyes are _field names.
Curently any nicos value that can be converted to float value is
converted and stored in additional _field with name field_float. This is
done for the sake of compatibility with external tools which might
process data types obtained from the DB. Nicos cache still operates with
values stored as strings.
This patch can be run and tested in demo, just need to take care of the
demo cache settings. An example of such settings left as commented code
in cache.py.
A script to transfer existing nicos cache to InfluxDB2 instance is
provided.

Fixes: #4528
Change-Id: I3803a2cff0689d2c8ee5a1a232625693ac0e18e0
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/29796
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>