core: history queries pass interval for data points
This is very useful for time series databases, such as InfluxDB,
because it reduces number of data points the database should return.
This increases performance of history queries quite significantly.

When history is queried the communication message has empty value
field. This value field is used to pass the desired interval. When
the interval is not passed, the corresponding filter command is not
applied at the cache back-end.
Currently unused for flat-file and kafka cache back-ends.

Change-Id: Ie5137bd281357c595be55abf90fac72a4ef4ca47
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/29838
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>