cache: updated influxdb-transfer
Fixed mistake for copy tool:
At line 240 there was a mistake, obviously value should not be changed,
but value_float.
When cache is copied 'None', '[]', '{}', '()' values should be copied
as well.
Copying is done in parallel python processes to speed up the routine.
After copying process is complete, any errors occurred will be printed.
The errors are stored in InfluxDB in a separate bucket.

Added tool to read errors which occurred during copying the cache and
output them to the command line.

Added tool to compare values from two independent nicos cache services,
currently is assumed that cache1 is bound to port 14869, and cache2 -
to 14870. This might be changed later.
Due to queries might return lot of data, the comparison occurs day by
day. Process of comparison is stored in InfluxDB, it stores key/subkey
and a timestamp of a day, in which comparison between two caches
matches.

Added automated tool, that parses the InfluxDB for keys and uses
previous tool to compare values and generate output for user.
This is done in multiple python processes to speed up the routine.

Change-Id: If378c5d006a7009c4a0eef02ca2666fde270ffc2
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/31085
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>