DataSinks: Add automatic single raw image file reader
The information about the image size is given either at the end of the
single raw file or in the corresponding header file. The routine to
parse the image information is in both cases the same as well as the
image reading routine.

The ImageReader class itself may decide which of the strategy is used
depending on the existence of the separate header file.

Change-Id: Ic0f6e6f5c23b8c52e21a8e3cf3bdd4bef04ee5f7
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/31548
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>