influxdb cache: improvements
It is recommended to use ciso8601 with influxdb-client for parsing
dates. ciso8601 is written in C and is much faster than built-in
Python datetime. Using this module speeds up processing huge queries
by a factor of ~6. Now history queries are faster than the ones
in flat file database.

Csv field size is set to 10 MB to handle queries with big fields,
such as scripts.

Influxdb-client timeout is set to 30 s.

There are 2 query methods in the wrapper now: the one which queries
last values since the beginning of nicos times, and the other that
queries the history.

History queries now yield records via query_stream() method.

Change-Id: I3d5e3ac0f314cd1b45d1e4e05088a9a879132503
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/31168
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>