Axis: Add method to set the motor position
The code for setting the position of the motor was in several lines
used. This didn't allow to have a single point for skipping the
re-adjusting the motor positions to coder or observer readings.

There is a request to skip this setting in a special axis class for the
Refsans instrument.

Change-Id: I85bce8d6f7be63cff64c1be0e5a91ace50fad7ae
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/33459
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>